fan speed can be 16 bit

+++ main/EMC2101.h
@@ -58,7 +58,7 @@ typedef enum {
 
 void EMC2101_set_fan_speed(float);
 void EMC2101_read(void);
-uint32_t EMC2101_get_fan_speed(void);
+uint16_t EMC2101_get_fan_speed(void);
 void EMC2101_set_config(uint8_t);
 
 #endif /* EMC2101_H_ */
